conda_requirements.txt
ansi2html @ file:///Users/runner/miniforge3/conda-bld/ansi2html_1695769044760/work
attrs==23.1.0
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1698890160476/work
boto3==1.33.12
botocore==1.33.12
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1695989934239/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1700303426725/work/certifi
cfgv==3.4.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1692311806742/work
dash @ file:///home/conda/feedstock_root/build_artifacts/dash_1698382724914/work
dash-auth==2.0.0
dash-bootstrap-components==1.5.0
dash-bootstrap-templates==1.1.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-mantine-components==0.12.0
dash-table==5.0.0
dash-tools==1.12.0
distlib==0.3.8
dnspython==2.4.2
docopt==0.6.2
filelock==3.13.1
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1696349106907/work
gunicorn==21.2.0
identify==2.5.33
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
iniconfig==2.0.0
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1648147185463/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1654302431367/work
jmespath==1.0.1
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1695367660391/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1697083700168/work
nodeenv==1.8.0
numpy==1.26.2
packaging==21.3
pandas==2.1.3
Pillow==10.1.0
pipreqs==0.4.12
platformdirs==4.1.0
plotly @ file:///home/conda/feedstock_root/build_artifacts/plotly_1698272730927/work
pluggy==1.3.0
pre-commit==2.16.0
py==1.11.0
pymongo==4.6.1
pyparsing==3.1.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pytest==6.2.5
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
rdkit==2023.9.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
retrying==1.3.3
ruamel.yaml==0.17.21
s3transfer==0.8.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1692026804430/work
termcolor==1.1.0
toml==0.10.2
tqdm==4.66.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1695040754690/work
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.25.0
visdcc==0.0.50
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1698235201373/work
xarray==2023.12.0
yarg==0.1.9
yolk3k==0.9
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
